% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_eb.R
\name{ebal_torch}
\alias{ebal_torch}
\title{ebal implementation with autodiff via torch}
\usage{
ebal_torch(X0, X1, base_weight = NULL, maxit = 200)
}
\arguments{
\item{X0}{donor units matrix}

\item{X1}{target moments}

\item{base_weight}{base weight vector}
}
\value{
list with weights and coefficients
}
\description{
ebal implementation with autodiff via torch
}
