% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.R
\name{getsquares}
\alias{getsquares}
\title{Generate Matrix of Squared Terms}
\usage{
getsquares(mat)
}
\arguments{
\item{mat}{n by k numeric matrix of covariates.}
}
\value{
matrix with additional columns
}
\description{
Takes a matrix of covariates and generates a new matrix
that contains the original covariates and all squared terms. Squared
terms for binary covariates are omitted.
}
