% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltest.collect.R
\name{baltest.collect}
\alias{baltest.collect}
\title{Collect Covariate Balance Statistics}
\usage{
baltest.collect(matchbal.out, var.names, after = TRUE)
}
\arguments{
\item{var.names}{A vector of covariate names}

\item{after}{A logical flag for whether the results from before or after Matching should be summarized. If TRUE baltest.collect summarizes the results from the covariate balance checks that MatchBalance computes in the matched data. If FALSE the results from the balance checks in the unmatched data are used.}

\item{matchbal}{An object from a call to MatchBalance(Matching)}
}
\value{
vector of treatment effects
}
\description{
A function that summarizes the covariate balance statistics that are computed by MatchBalance(Matching) in a balance table.
}
