% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{eb}
\alias{eb}
\title{internal function to compute entropy balancing weights}
\usage{
eb(
  tr.total = tr.total,
  co.x = co.x,
  coefs = coefs,
  base.weight = base.weight,
  max.iterations = max.iterations,
  constraint.tolerance = constraint.tolerance,
  print.level = print.level
)
}
\arguments{
\item{tr.total}{treatment moments}

\item{co.x}{covariate matrix}

\item{coefs}{starting coefs}

\item{base.weight}{base weight}

\item{max.iterations}{max number of iterations}

\item{constraint.tolerance}{tolerance for constraint}

\item{print.level}{print level for debugging}
}
\value{
list with results
}
\description{
internal function to compute entropy balancing weights
}
