% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.R
\name{matrixmaker}
\alias{matrixmaker}
\title{Generate Matrix of One-way Interactions and Squared Terms}
\usage{
matrixmaker(mat)
}
\arguments{
\item{mat}{n by k numeric matrix of covariates.}
}
\value{
matrix with additional columns
}
\description{
Takes a matrix of covariates and generates a new matrix
that contains the original covariates, all one-way interaction terms,
and all squared terms.
}
